Cleanup models, controllers and patches

+++ app/controllers/mail_reminders_controller.rb
@@ -1,8 +1,6 @@
 class MailRemindersController < ApplicationController
-  unloadable
-
   before_filter :find_project
-  before_filter :authorize, :only => :index
+  before_filter :authorize, only: :index
 
   def index
     needs_refresh = false
@@ -18,9 +16,9 @@ class MailRemindersController < ApplicationController
   end
 
   def create
-    parameters = ActionController::Parameters.new(params[:reminder])
-    reminder = MailReminder.new(parameters.permit(:project_id, :query_id, :interval))
+    reminder = MailReminder.new(reminder_params)
     reminder.interval_value = params[:interval_value].to_i
+
     if reminder.save
       Role.find_all_givable.each do |role|
         if params[role.name.downcase]
@@ -31,9 +29,9 @@ class MailRemindersController < ApplicationController
         end
       end
 
-      flash[:notice] = t :reminder_created
+      flash[:notice] = t(:reminder_created)
     else
-      flash[:error] = t :reminder_not_created
+      flash[:error] = t(:reminder_not_created)
     end
 
     render partial: 'reload'
@@ -41,7 +39,8 @@ class MailRemindersController < ApplicationController
 
   def update
     reminder = MailReminder.find(params[:id])
-    if request.put? && reminder.update_attributes(params[:reminder])
+
+    if request.put? && reminder.update_attributes(reminder_params)
       reminder.interval_value = params[:interval_value]
       Role.find_all_givable.each do |role|
         if reminder.roles.include?(role) && params[role.name.downcase].nil?
@@ -62,15 +61,14 @@ class MailRemindersController < ApplicationController
 
   def destroy
     reminder = MailReminder.find(params[:id])
-    if reminder
-      reminder.destroy
-    end
+    reminder.destroy if reminder
 
     render partial: 'reload'
   end
 
   def update_interval_values
     vals = MailReminder.interval_values_for(params[:interval])
+
     begin
       reminder = MailReminder.find(params[:mail_reminder_id])
     rescue ActiveRecord::RecordNotFound
@@ -88,5 +86,9 @@ class MailRemindersController < ApplicationController
     rescue ActiveRecord::RecordNotFound
       @project = Project.find(params[:reminder][:project_id]) if params[:reminder]
     end
+  end
+
+  def reminder_params
+    params.require(:reminder).permit(:project_id, :query_id, :interval)
   end
 end
